sharp_module/requirements.txt
aioftp==0.22.3
aiohttp==3.9.3
aiosignal==1.3.1
aniso8601==9.0.1
asciitree==0.3.3
asdf==3.1.0
asdf-astropy==0.6.0
asdf_coordinates_schemas==0.3.0
asdf_standard==1.1.1
asdf_transform_schemas==0.5.0
asdf_unit_schemas==0.2.0
astropy==5.1
astropy-healpix==1.0.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
cdflib==1.2.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.0
cycler==0.12.1
dask==2024.3.1
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
drms==0.6.4
executing==2.0.1
fasteners==0.19
fastjsonschema==2.19.1
Flask==3.0.2
flask-restx==1.3.0
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
Glymur==0.12.9.post1
greenlet==3.0.3
h5netcdf==1.3.0
h5py==3.10.0
idna==3.6
imageio==2.34.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipython==8.12.3
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
lazy_loader==0.3
locket==1.0.0
lxml==5.1.0
MarkupSafe==2.1.5
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==3.0.2
mpl-animators==1.1.1
multidict==6.0.5
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
networkx==3.2.1
numcodecs==0.12.1
numpy==1.23.5
packaging==24.0
pandas==1.4.4
pandocfilters==1.5.1
parfive==2.0.2
parso==0.8.3
partd==1.4.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pipreqs==0.5.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyerfa==2.0.1.1
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.34.0
reproject==0.13.0
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.18.0
scikit-image==0.22.0
scipy==1.12.0
semantic-version==2.10.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.29
stack-data==0.6.3
sunpy==5.0.0
tifffile==2024.2.12
tinycss2==1.2.1
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.1
yarg==0.1.9
yarl==1.9.4
zarr==2.17.1
zeep==4.2.1
zipp==3.18.1
